webkitgtk-2.20.5

+++ Tools/WebKitTestRunner/wpe/EventSenderProxyWPE.cpp
@@ -27,10 +27,10 @@
 #include "EventSenderProxy.h"
 
 #include "HeadlessViewBackend.h"
-#include "NotImplemented.h"
 #include "PlatformWebView.h"
 #include "TestController.h"
-#include <wpe/view-backend.h>
+#include <WebCore/NotImplemented.h>
+#include <wpe/wpe.h>
 #include <xkbcommon/xkbcommon-keysyms.h>
 #include <xkbcommon/xkbcommon.h>
 
@@ -289,7 +289,7 @@ void EventSenderProxy::addTouchPoint(int x, int y)
 
 void EventSenderProxy::updateTouchPoint(int index, int x, int y)
 {
-    ASSERT(index >= 0 && index <= m_touchEvents.size());
+    ASSERT(index >= 0 && static_cast<size_t>(index) <= m_touchEvents.size());
 
     auto& rawEvent = m_touchEvents[index];
     rawEvent.x = x;
@@ -364,7 +364,7 @@ void EventSenderProxy::clearTouchPoints()
 
 void EventSenderProxy::releaseTouchPoint(int index)
 {
-    ASSERT(index >= 0 && index <= m_touchEvents.size());
+    ASSERT(index >= 0 && static_cast<size_t>(index) <= m_touchEvents.size());
 
     auto& rawEvent = m_touchEvents[index];
     rawEvent.time = m_time;
